Refractor Interface
Change utility test interface to class, specificities of interface were not use
